+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 ============================
+# 2.0.0 (Nov 2, 2023)
 
+- Moved source code to `src` folder
+- Added `davidhirtz\yii2\media\video\components\helpers\Video` HTML video helper tag
+- Locked `davidhirtz/yii2-media` to version `2.0`
+ 
 # 1.0.3 (Nov 2, 2023)
 
 - Locked `davidhirtz/yii2-media` to version `^1.3`, upgrade to version 2.0 to use the latest version of this package